feat: migrate to Rust API, drop deprecated Python API symlinks, fix all lint warnings
- Remove stale `frontend` and `reference` root symlinks (both pointed to code
  now properly located under clients/frontend and services/api-py-deprecated)
- Add Makefile targets: python-lint, python-fmt, python-test, frontend-lint,
  frontend-fmt, and aggregate lint/fmt/test targets
- Rust API (services/api): complete provider, scraper, route, and stream
  rewrites; drop individual provider files (alldebrid, debridlink, easydebrid,
  offcloud, pikpak, premiumize, realdebrid, seedr, stremthru, torbox) now
  consolidated into providers/torrents and providers/usenet modules
- Fix all cargo clippy -D warnings: replace invalid look-around regex patterns
  with word-boundary equivalents, remove redundant let bindings, replace manual
  min/max chains with clamp(), use strip_prefix(), simplify boolean expressions,
  and add #[allow] annotations for structural too-many-arguments and
  type-complexity cases
- Run cargo fmt and ruff format; fix two unused imports in migration scripts
- tests/conftest.py: register services/api-py-deprecated as the `reference`
  package so existing unit tests continue to resolve imports without the symlink
- Update db models, migrations, docker-compose configs, and env-reference docs
  to reflect Postgres-only schema and new Rust service layout

+++ services/api/src/routes/content/streams.rs
@@ -20,7 +20,7 @@ use crate::state::AppState;
 
 // ─── Auth helpers ─────────────────────────────────────────────────────────────
 
-fn validate_moderator_token(headers: &HeaderMap, secret_key: &str) -> Option<i64> {
+fn validate_moderator_token(headers: &HeaderMap, secret_key: &str) -> Option<i32> {
     let token = headers
         .get("authorization")
         .and_then(|v| v.to_str().ok())
@@ -61,7 +61,7 @@ fn validate_moderator_token(headers: &HeaderMap, secret_key: &str) -> Option<i64
 pub async fn delete_stream(
     State(state): State<Arc<AppState>>,
     headers: HeaderMap,
-    Path(stream_id): Path<i64>,
+    Path(stream_id): Path<i32>,
 ) -> Response {
     // 1. Validate moderator token
     let _user_id = match validate_moderator_token(&headers, &state.config.secret_key_raw) {
@@ -75,7 +75,7 @@ pub async fn delete_stream(
         }
     };
 
-    let stream_id_i32 = stream_id as i32;
+    let stream_id_i32 = stream_id;
 
     // 2. Check stream exists and get stream_type
     let stream_type: Option<String> =
